+++ module/Mailing/src/Mailing/Controller/MailingController.php
@@ -4,29 +4,45 @@ namespace Mailing\Controller;
 
 use Mailing\Model\Db\MailType;
 use Mailing\Service\MailType\MailTypeServiceAwareTrait;
+use UnicaenEtat\Service\Etat\EtatServiceAwareTrait;
+use UnicaenEtat\Service\EtatType\EtatTypeServiceAwareTrait;
 use UnicaenUtilisateur\Service\User\UserServiceAwareTrait;
 use Mailing\Model\Db\Mail;
 use Mailing\Service\Mailing\MailingServiceAwareTrait;
 use Zend\Mvc\Controller\AbstractActionController;
+use Zend\View\Model\JsonModel;
 use Zend\View\Model\ViewModel;
 
 class MailingController extends AbstractActionController {
     use MailingServiceAwareTrait;
     use MailTypeServiceAwareTrait;
+    use EtatServiceAwareTrait;
+    use EtatTypeServiceAwareTrait;
     use UserServiceAwareTrait;
 
     public function indexAction()
     {
+        $fromQueries  = $this->params()->fromQuery();
+        $adresse      = (trim($fromQueries['adresse']['label']) != '')?trim($fromQueries['adresse']['label']):null;
+        $etatId       = $fromQueries['etat'];
+        $typeId       = $fromQueries['type'];
+
+        $params = ['adresse' => $adresse, 'etat' => $etatId, 'type' => $typeId];
+
         /**
          * @var Mail[] $mails
-         * @var MailType[] $types
+         * @var MailType[] $types
          */
-        $mails = $this->getMailingService()->getMails();
+        $mails = $this->getMailingService()->getMailsWithFiltre($params);
         $types = $this->getMailTypeService()->getMailsTypes();
+        $type = $this->getEtatTypeService()->getEtatTypeByCode('MAIL');
+        $etats = $this->getEtatService()->getEtatsByType($type);
 
         return new ViewModel([
             'mails' => $mails,
             'types' => $types,
+            'etats' => $etats,
+            'params' => $params,
         ]);
     }
 
@@ -70,5 +86,15 @@ class MailingController extends AbstractActionController {
 
         if ($retour !== null) return $this->redirect()->toUrl($retour);
         return $this->redirect()->toRoute('mailing');
+    }
+
+    public function rechercherAdresseAction()
+    {
+        if (($term = $this->params()->fromQuery('term'))) {
+            $adresses = $this->getMailingService()->findAdressetByTerm($term);
+            $result = $this->getMailingService()->formatAdresseJSON($adresses);
+            return new JsonModel($result);
+        }
+        exit;
     }
 }